Re-organize the history decorator.
