+++ resources/assets/helpers/createHistoryHashObserver.js
